+++ backend-pages/all-users-page-upgrades.php
@@ -32,37 +32,98 @@ function ssv_frontend_members_users_page()
 
 add_action('admin_menu', 'ssv_add_ssv_frontend_members_users_page');
 
-function add_course_section_filter()
+function ssv_custom_user_column_values($val, $column_name, $user_id)
 {
-    $meta_key   = isset($_SESSION['meta_key']) ? $_SESSION['meta_key'] : '';
-    $meta_value = isset($_SESSION['meta_value']) ? $_SESSION['meta_value'] : '';
-    if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-        if (isset($_POST['clear_filters']) || !isset($_POST['meta_value']) || empty($_POST['meta_value'])) {
-            unset($_SESSION['meta_key']);
-            unset($_SESSION['meta_value']);
-            $meta_key = '';
-            $meta_value = '';
+    $frontendMember = FrontendMember::get_by_id($user_id);
+    if ($column_name == 'ssv_member') {
+        $username_block = '';
+        $username_block .= '<img style="float: left; margin-right: 10px; margin-top: 1px;" class="avatar avatar-32 photo" src="' . esc_url($frontendMember->getMeta('profile_picture')) . '" height="32" width="32"/>';
+        $username_block .= '<strong>' . $frontendMember->getProfileLink('_blank') . '</strong><br/>';
+        $directDebitPDF  = $frontendMember->getProfileURL() . '&view=directDebitPDF';
+        $editURL         = 'user-edit.php?user_id=' . $frontendMember->ID . '&wp_http_referer=%2Fwp-admin%2Fusers.php';
+        $capebilitiesURL = 'users.php?page=users-user-role-editor.php&object=user&user_id=' . $frontendMember->ID;
+        $username_block .= '<div class="row-actions"><span class="direct_debit_pdf"><a href="' . esc_url($directDebitPDF) . '" target="_blank">PDF</a> | </span><span class="edit"><a href="' . esc_url($editURL) . '">Edit</a> | </span><span class="capabilities"><a href="' . esc_url($capebilitiesURL) . '">Capabilities</a></span></div>';
+        return $username_block;
+    } elseif (ssv_starts_with($column_name, 'ssv_')) {
+        return $frontendMember->getMeta(str_replace('ssv_', '', $column_name));
+    }
+    return $val;
+}
+
+add_filter('manage_users_custom_column', 'ssv_custom_user_column_values', 10, 3);
+
+function ssv_custom_user_columns($column_headers)
+{
+    unset($column_headers);
+    $column_headers['cb'] = '<input type="checkbox" />';
+    global $wpdb;
+    if (get_option('ssv_frontend_members_main_column') == 'wordpress_default') {
+        $column_headers['username'] = 'Username';
+    } else {
+        $url = $_SERVER['REQUEST_URI'];
+        if (empty($_GET)) {
+            $url .= '?orderby=name';
+        } elseif (!isset($_GET['orderby'])) {
+            $url .= '&orderby=name';
+        } elseif (!isset($_GET['order'])) {
+            $url .= '&order=DESC';
+        } elseif ($_GET['order'] == 'DESC') {
+            $url .= '&order=ASC';
         } else {
-            $_SESSION['meta_key'] = $meta_key = $_POST['meta_key'];
-            $_SESSION['meta_value'] = $meta_value = $_POST['meta_value'];
+            $url .= '&order=DESC';
+        }
+        $column_headers['ssv_member'] = '<a href="' . $url . '">Member</a>';
+    }
+    $selected_columns = json_decode(get_option('ssv_frontend_members_user_columns'));
+    $selected_columns = $selected_columns ?: array();
+    foreach ($selected_columns as $column) {
+        $sql   = 'SELECT field_id FROM ' . FRONTEND_MEMBERS_FIELD_META_TABLE_NAME . ' WHERE meta_key = "name" AND meta_value = "' . $column . '"';
+        $sql   = 'SELECT field_title FROM ' . FRONTEND_MEMBERS_FIELDS_TABLE_NAME . ' WHERE id = (' . $sql . ')';
+        $title = $wpdb->get_var($sql);
+        if (ssv_starts_with($column, 'wp_')) {
+            $column                              = str_replace('wp_', '', $column);
+            $column_headers[strtolower($column)] = $column;
+        } else {
+            $column_headers['ssv_' . $column] = $title;
         }
     }
-    $options = array('' => 'Select Key');
+    return $column_headers;
+}
+
+add_action('manage_users_columns', 'ssv_custom_user_columns');
+
+function add_course_section_filter()
+{
     $fields = FrontendMembersField::getAll(array('registration_page' => 'no', 'field_type' => 'input'));
+    if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+        foreach ($fields as $field) {
+            /** @var FrontendMembersFieldInput $field */
+            if (isset($_POST['clear_filters']) || !isset($_POST['filter_' . $field->name]) || empty($_POST['filter_' . $field->name])) {
+                unset($_SESSION['filter_' . $field->name]);
+            } else {
+                $_SESSION['filter_' . $field->name] = $_POST['filter_' . $field->name];
+            }
+        }
+    }
+    $filters  = '<h1>Filters</h1>';
+    $selected = json_decode(get_option('ssv_frontend_members_user_filters'));
+    $selected = $selected ?: array();
     foreach ($fields as $field) {
         /** @var FrontendMembersFieldInput $field */
-        $options[$field->name] = $field->title;
+        if (in_array($field->name, $selected)) {
+            $filters .= '<div style="display: inline-block; margin-right: 10px;">';
+            $filters .= $field->getFilter();
+            $filters .= '</div>';
+        }
     }
-    $filters = ssv_get_td(ssv_get_select('Meta Key', null, $meta_key, $options, array(), false, null, false));
-    $filters .= ssv_get_td('<input type="text" id="meta_value" name="meta_value" value="'.$meta_value.'">');
-    $filters .= ssv_get_td('<button type="submit" name="submit">Filter</button>');
-    $filters .= ssv_get_td('<button type="submit" name="clear_filters">Clear Filters</button>');
+    $filters .= '<br/>';
+    $filters .= '<button type="submit" value="submit" class="button">Filter</button>';
     ?>
     <script>
         window.onload = function () {
             jQuery(document).ready(function ($) {
                 var old_filter_area = $('.subsubsub');
-                old_filter_area.after('<form name="filter_form" method="post"><table id="filter_area"></table></form>');
+                old_filter_area.after('<form name="filter_form" method="post"><div id="filter_area"></div></form>');
                 old_filter_area.remove();
                 var filter_area = $('#filter_area');
                 filter_area.html('<?php echo $filters; ?>');
@@ -79,12 +140,19 @@ function filter_users_by_course_section($query)
 {
 //    die('<br/><br/>test');
     global $wpdb;
-    if (strpos($_SERVER['REQUEST_URI'], 'users.php') !== false && isset($_SESSION["meta_value"])) {
-        $filters = array($_SESSION["meta_key"] => $_SESSION["meta_value"]);
+    if (strpos($_SERVER['REQUEST_URI'], 'users.php') !== false) {
+        $filters = array();
+        $fields = FrontendMembersField::getAll(array('registration_page' => 'no', 'field_type' => 'input'));
+        foreach ($fields as $field) {
+            /** @var FrontendMembersFieldInput $field */
+            if (isset($_SESSION['filter_' . $field->name])) {
+                $filters[$field->name] = $_SESSION['filter_' . $field->name];
+            }
+        }
         foreach ($filters as $filter => $value) {
             $table_alias = $filter . 'meta';
             $query->query_from .= " JOIN {$wpdb->usermeta} {$table_alias} ON {$table_alias}.user_id = {$wpdb->users}.ID AND {$table_alias}.meta_key = '{$filter}'";
-            $query->query_where = "WHERE {$table_alias}.meta_value LIKE '%{$value}%'";
+            $query->query_where .= " AND {$table_alias}.meta_value LIKE '%{$value}%'";
         }
     }
 }
